Server/requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
# python<3.9
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.12
